requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array_record==0.5.1
arrow==1.3.0
astroid==3.3.8
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
autopep8==2.3.2
Babel==2.15.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
comm==0.1.4
command-not-found==0.3
confection==0.1.4
contourpy==1.2.1
cryptography==3.4.8
cycler==0.12.1
cymem==2.0.8
dbus-python==1.2.18
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dialogflow==2.0.0
dill==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.1
dm-tree==0.1.8
etils==1.7.0
exceptiongroup==1.2.1
executing==1.2.0
fastai==2.7.15
fastcore==1.5.32
fastdownload==0.0.7
fastjsonschema==2.19.1
fastprogress==1.0.3
filelock==3.14.0
flake8==7.1.2
Flask==3.0.2
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
fsspec==2024.3.1
gast==0.5.4
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud==0.34.0
google-cloud-dialogflow==2.28.1
google-cloud-dialogflow-cx==1.31.1
google-cloud-vision==3.7.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
graphviz==0.20.1
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
idna==3.6
imageio==2.34.1
importlib-metadata==4.6.4
importlib_resources==6.4.0
iniconfig==2.0.0
ipykernel==6.25.0
ipython==8.32.0
ipywidgets==8.1.2
isoduration==20.11.0
isort==6.0.1
itsdangerous==2.1.2
jedi==0.19.0
jeepney==0.7.1
Jinja2==3.1.3
joblib==1.4.0
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.10
keras==3.9.0
keyring==23.5.0
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
libclang==16.0.6
llvmlite==0.44.0
marisa-trie==1.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
more-itertools==8.10.0
mpmath==1.3.0
murmurhash==1.0.10
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.5.7
netifaces==0.11.0
networkx==3.3
notebook==7.2.0
notebook_shim==0.2.4
numba==0.61.0
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.5.3.2
nvidia-cuda-cupti-cu12==12.5.82
nvidia-cuda-nvcc-cu12==12.5.82
nvidia-cuda-nvrtc-cu12==12.5.82
nvidia-cuda-runtime-cu12==12.5.82
nvidia-cudnn-cu11==8.6.0.163
nvidia-cudnn-cu12==9.3.0.75
nvidia-cufft-cu12==11.2.3.61
nvidia-curand-cu12==10.3.6.82
nvidia-cusolver-cu12==11.6.3.83
nvidia-cusparse-cu12==12.5.1.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
opt-einsum==3.3.0
optree==0.12.1
overrides==7.7.0
packaging==23.2
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==3.10.0
pluggy==1.5.0
pooch==1.8.1
preshed==3.0.9
prometheus_client==0.20.0
promise==2.3
prompt_toolkit==3.0.50
proto-plus==1.23.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydot==2.0.0
pyflakes==3.2.0
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pylint==3.3.4
pyparsing==2.4.7
pytest==8.3.4
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-json-logger==2.0.7
pytorch-tabnet==4.1.0
pytz==2024.1
PyYAML==5.4.1
pyzmq==25.1.0
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rpds-py==0.18.1
rsa==4.9
scikit-image==0.23.2
scikit-learn==1.6.1
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.3
shap==0.46.0
six==1.16.0
slicer==0.0.8
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.2
sympy==1.12
systemd-python==234
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
tensorflow-metadata==1.15.0
tensorflow-probability==0.24.0
termcolor==2.4.0
terminado==0.18.1
tf-explain==0.3.1
thinc==8.2.3
threadpoolctl==3.5.0
tifffile==2024.4.24
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.13.2
torch==2.3.0
torchaudio==2.3.0+cu118
torchvision==0.18.0
tornado==6.3.2
tqdm==4.66.2
traitlets==5.14.3
triton==2.3.0
typer==0.9.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
wadllib==1.3.6
wasabi==1.1.2
wcwidth==0.2.6
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.1
widgetsnbextension==4.0.10
wrapt==1.14.1
xgboost==2.1.4
yapf==0.43.0
zipp==1.0.0
